// src/views/Settings/PaymentSettings/types.ts
export interface FormData {
  companyName: string;
  invoiceTitle: string;
  thankYouMessage: string;
  questionsMessage: string;
  closingMessage: string;
  regards: string;
  businessName: string;
  logoUrl: string;
  invoiceNumber: string;
  invoiceDate: string;
  amountDue: string;
  dueDate: string;
}